Add Windows support.
ava/cli.js makes use of shebang string, which fails on windows.
A solution would be to run cli.js via node. To do so execFile in grunt-ava
calls node and supplies it with args containing cli.js as the first element.

Solution has been successfully tested on windows, mac and linux.